# docs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Cellar",
    "description": "Simple secret sharing with the infrastructure you already trust",
    "contact": {
      "name": "Aria Vesta",
      "url": "http://cellar-app.io",
      "email": "dev@ariavesta.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://gitlab.com/cellar-app/cellar-api/-/blob/main/LICENSE.txt"
    }
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "paths": {
    "/health-check": {
      "get": {
        "tags": [
          "common"
        ],
        "summary": "Health Check",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/models.HealthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/secrets": {
      "post": {
        "tags": [
          "v1"
        ],
        "summary": "Create Secret",
        "requestBody": {
          "description": "Add secret",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/models.CreateSecretRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/models.SecretMetadataResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/httputil.HTTPError"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/httputil.HTTPError"
                }
              }
            }
          }
        },
        "x-codegen-request-body-name": "secret"
      }
    },
    "/v1/secrets/{id}": {
      "get": {
        "tags": [
          "v1"
        ],
        "summary": "Get Secret Metadata",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Secret ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/models.SecretMetadataResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/httputil.HTTPError"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/httputil.HTTPError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "v1"
        ],
        "summary": "Delete Secret",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Secret ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "",
            "content": {}
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/httputil.HTTPError"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/httputil.HTTPError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/secrets/{id}/access": {
      "post": {
        "tags": [
          "v1"
        ],
        "summary": "Access Secret Content",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Secret ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/models.SecretContentResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/httputil.HTTPError"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/httputil.HTTPError"
                }
              }
            }
          }
        }
      }
    },
    "/v2/secrets": {
      "post": {
        "tags": [
          "v2"
        ],
        "summary": "Create Secret",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "required": [
                  "expiration_epoch"
                ],
                "type": "object",
                "properties": {
                  "content": {
                    "type": "string",
                    "description": "Secret content"
                  },
                  "access_limit": {
                    "type": "integer",
                    "description": "Access limit",
                    "format": "int32"
                  },
                  "expiration_epoch": {
                    "type": "integer",
                    "description": "Expiration of the secret in Unix Epoch Time",
                    "format": "int32"
                  },
                  "file": {
                    "type": "string",
                    "description": "Secret content as a file",
                    "format": "binary"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/models.SecretMetadataResponseV2"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/httputil.HTTPError"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/httputil.HTTPError"
                }
              }
            }
          }
        }
      }
    },
    "/v2/secrets/{id}": {
      "get": {
        "tags": [
          "v2"
        ],
        "summary": "Get Secret Metadata",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Secret ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/models.SecretMetadataResponseV2"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/httputil.HTTPError"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/httputil.HTTPError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "v2"
        ],
        "summary": "Delete Secret",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Secret ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "",
            "content": {}
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/httputil.HTTPError"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/httputil.HTTPError"
                }
              }
            }
          }
        }
      }
    },
    "/v2/secrets/{id}/access": {
      "post": {
        "tags": [
          "v2"
        ],
        "summary": "Access Secret Content. If the content is a file it the response will be an application/octet-stream",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Secret ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/models.SecretContentResponse"
                }
              },
              "application/octet-stream": {
                "schema": {
                  "$ref": "#/components/schemas/models.SecretContentResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/httputil.HTTPError"
                }
              },
              "application/octet-stream": {
                "schema": {
                  "$ref": "#/components/schemas/httputil.HTTPError"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/httputil.HTTPError"
                }
              },
              "application/octet-stream": {
                "schema": {
                  "$ref": "#/components/schemas/httputil.HTTPError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "httputil.HTTPError": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "example": 400
          },
          "message": {
            "type": "string",
            "example": "status bad request"
          }
        }
      },
      "models.CreateSecretRequest": {
        "type": "object",
        "properties": {
          "access_limit": {
            "type": "integer",
            "example": 10
          },
          "content": {
            "type": "string",
            "example": "my very secret text"
          },
          "expiration_epoch": {
            "type": "integer",
            "example": 1577836800
          }
        }
      },
      "models.Health": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Redis"
          },
          "status": {
            "type": "string",
            "example": "healthy"
          },
          "version": {
            "type": "string",
            "example": "1.0.0"
          }
        }
      },
      "models.HealthResponse": {
        "type": "object",
        "properties": {
          "datastore": {
            "$ref": "#/components/schemas/models.Health"
          },
          "encryption": {
            "$ref": "#/components/schemas/models.Health"
          },
          "host": {
            "type": "string",
            "example": "localhost"
          },
          "status": {
            "type": "string",
            "example": "healthy"
          },
          "version": {
            "type": "string",
            "example": "1.0.0"
          }
        }
      },
      "models.SecretContentResponse": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string",
            "example": "my very secret text"
          },
          "id": {
            "type": "string",
            "example": "22b6fff1be15d1fd54b7b8ec6ad22e80e66275195c914c4b0f9652248a498680"
          }
        }
      },
      "models.SecretMetadataResponse": {
        "type": "object",
        "properties": {
          "access_count": {
            "type": "integer",
            "example": 1
          },
          "access_limit": {
            "type": "integer",
            "example": 10
          },
          "expiration": {
            "type": "string",
            "example": "1970-01-01 00:00:00 UTC"
          },
          "id": {
            "type": "string",
            "example": "22b6fff1be15d1fd54b7b8ec6ad22e80e66275195c914c4b0f9652248a498680"
          }
        }
      },
      "models.SecretMetadataResponseV2": {
        "type": "object",
        "properties": {
          "access_count": {
            "type": "integer",
            "example": 1
          },
          "access_limit": {
            "type": "integer",
            "example": 10
          },
          "content_type": {
            "type": "string",
            "example": "text"
          },
          "expiration": {
            "type": "string",
            "example": "1970-01-01 00:00:00 UTC"
          },
          "id": {
            "type": "string",
            "example": "22b6fff1be15d1fd54b7b8ec6ad22e80e66275195c914c4b0f9652248a498680"
          }
        }
      }
    }
  },
  "x-original-swagger-version": "2.0"
}